% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{download_file}
\alias{download_file}
\title{Downloads from Google Docs}
\usage{
download_file(
  file,
  gfile = NULL,
  path = "rmdrive",
  team_drive = NULL,
  restore_chunks = FALSE
)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).}

\item{team_drive}{character. The name of a Google Team Drive (optional).}

\item{restore_chunks}{logical value indicating whether to restore code chunks
in the document.}
}
\value{
`TRUE` if file from Google Drive was saved, `FALSE` otherwise
}
\description{
Downloads a text file from Google Drive and saves it as a local
  file if the local file does not exist or differs from file in Google Drive.
}
