% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{download_rmd}
\alias{download_rmd}
\title{Downloads raw rmarkdown from Google Docs}
\usage{
download_rmd(file, gfile)
}
\arguments{
\item{file}{The name (without file extension) of an .Rmd you want to upload to Google Docs}

\item{gfile}{The name (current or desired) of a Google Docs file (must be unique in your google docs account)}
}
\description{
Downloads a Google Docs file and changes the file extension to .Rmd
}
\examples{
download_rmd(file = "testingfile", gfile = "testingfile-1234")
}
