% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chunk.R
\name{extract_chunk}
\alias{extract_chunk}
\title{Extract chunks as list from document}
\usage{
extract_chunk(local_rmd)
}
\arguments{
\item{local_rmd}{character indicating the name of the file without extension}
}
\value{
TODO
}
\description{
Extract chunks as list from document
}
