% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{final_rmd}
\alias{final_rmd}
\title{Upload final compiled documemnt on Google Drive}
\usage{
final_rmd(file, path = "rmdrive", team_drive = NULL)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).
By default files are uploaded in the folder "rmdrive". To specify another
folder the full path is required (e.g., "rmdrive/my_folder"). Use
\code{NULL} to upload directly at the root level, although it is not
recommended.}

\item{team_drive}{character. The name of a Google Team Drive (optional).}
}
\description{
Render the final .Rdm file and upload the overall version to the specified folder
}
