% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chunk.R
\name{init_rmdrive}
\alias{init_rmdrive}
\title{Create .rmdrive folder if missing}
\usage{
init_rmdrive(local_rmd)
}
\arguments{
\item{local_rmd}{character indicating the name of the file without extension}
}
\description{
Create .rmdrive folder if missing
}
