% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chunk.R
\name{mkdir_rmdrive}
\alias{mkdir_rmdrive}
\title{Create .rmdrive folder if missing}
\usage{
mkdir_rmdrive(local_path)
}
\arguments{
\item{local_path}{string indicating the path where to create the folder}
}
\description{
Create .rmdrive folder if missing
}
