% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{render_file}
\alias{render_file}
\title{Render file from GoogleDrive}
\usage{
render_file(
  file,
  gfile = basename(file),
  path = "trackdown",
  team_drive = NULL,
  restore_chunks = FALSE
)
}
\arguments{
\item{file}{character. The path of a local `.Rmd` or `Rnw` file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).
By default files are uploaded in the folder "trackdown". To specify another
folder the full path is required (e.g., "trackdown/my_folder"). Use
\code{NULL} to upload directly at the root level, although it is not
recommended.}

\item{team_drive}{character. The name of a Google Team Drive (optional).}

\item{restore_chunks}{logical value indicating whether to restore code chunks
in the document.}
}
\description{
Renders file from GoogleDrive if there have been edits
}
