% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Sanitize Rmd file downloaded from GoogleDrive}
\usage{
stop_quietly(text = NULL)
}
\arguments{
\item{text}{a string indicating the message to display}
}
\value{
Sanitized file from Google Drive.
}
\description{
Adds a final EOL and removes double linebreaks added when downloading file
from GoogleDrive.
}
