% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown_auth.R
\name{trackdown_has_token}
\alias{trackdown_has_token}
\title{Is there a token on hand?}
\usage{
trackdown_has_token()
}
\value{
Logical.
}
\description{
Reports whether trackdown has stored a token, ready for use in downstream
requests.
}
\examples{
trackdown_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{trackdown_token}()}
}
\concept{low-level API functions}
