% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown_auth.R
\name{trackdown_token}
\alias{trackdown_token}
\title{Produce configured token}
\usage{
trackdown_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the trackdown API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=trackdown_auth]{trackdown_auth()}} is what they need. If there is no current
token, \code{\link[=trackdown_auth]{trackdown_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=trackdown_deauth]{trackdown_deauth()}}, \code{trackdown_token()}
returns \code{NULL}.
}
\examples{
\dontrun{
trackdown_token()
}

}
\seealso{
Other low-level API functions: 
\code{\link{trackdown_has_token}()}
}
\concept{low-level API functions}
