% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown_auth.R
\name{trackdown_user}
\alias{trackdown_user}
\title{Get info on current user}
\usage{
trackdown_user()
}
\value{
An email address or, if no token has been loaded, `NULL`.
}
\description{
Reveals the email address of the user associated with the current token.
If no token has been loaded yet, this function does not initiate auth.
}
\examples{
\dontrun{
trackdown_user()
}

}
\seealso{
[gargle::token_userinfo()], [gargle::token_email()],
[gargle::token_tokeninfo()]
}
