% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{update_rmd}
\alias{update_rmd}
\title{Updates \code{.Rmd} file in Google Drive}
\usage{
update_rmd(file, gfile = basename(file), path = NULL,
  team_drive = NULL)
}
\arguments{
\item{file}{character. The name (without file extension) of a local \code{.Rmd}
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).}

\item{team_drive}{character. The name of a Google Team Drive (optional).}
}
\description{
Replaces the content of an existing file in Google Drive with the contents of
a local \code{.Rmd} file.

\emph{Use with caution as tracked changes in the Google Drive file will be lost!}
}
