% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{upload_file}
\alias{upload_file}
\title{Upload file to Google Drive for collaborative editing}
\usage{
upload_file(
  file,
  gfile = basename(file),
  path = "rmdrive",
  team_drive = NULL,
  hide_chunks = FALSE,
  upload_report = FALSE
)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).
By default files are uploaded in the folder "rmdrive". To specify another
folder the full path is required (e.g., "rmdrive/my_folder"). Use
\code{NULL} to upload directly at the root level, although it is not
recommended.}

\item{team_drive}{character. The name of a Google Team Drive (optional).}

\item{hide_chunks}{logical value indicating whether to remove code chunks
from the text document. Placeholders of  type "[[chunck_<name>]]" are
displayed instead.}

\item{upload_report}{logical value indicating whether to upload an
additional pdf file with the chunks output (e.g., figures and tables). Note
that this require the time to compile the document.}
}
\description{
Uploads a local file to Google Drive as a plain text document. Will
only upload the file if it doesn't already exist in the chosen location. By
default files are uploaded in the folder "rmdrive", if is not available on
Google Drive, permission to create is required to the user. To update an
existing file \code{\link{update_file}}.
}
