% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{upload_rmd}
\alias{upload_rmd}
\title{Upload \code{.Rmd} file to Google Drive for collaborative editing}
\usage{
upload_rmd(file, gfile = basename(file), path = NULL,
  team_drive = NULL)
}
\arguments{
\item{file}{character. The name (without file extension) of a local \code{.Rmd}
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).}

\item{team_drive}{character. The name of a Google Team Drive (optional).}
}
\description{
Uploads a local \code{.Rmd} file to Google Drive as a plain text document.
Will only upload the file if it doesn't already exist in the chosen
location. To update an existing file \link{update_rmd}.
}
