% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{download_file}
\alias{download_file}
\title{Downloads from Google Docs}
\usage{
download_file(file, gfile = NULL, gpath = "trackdown", team_drive = NULL)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{gpath}{character. (Sub)directory in My Drive or a Team Drive (optional).}

\item{team_drive}{character. The name of a Google Team Drive (optional).}
}
\value{
`TRUE` if file from Google Drive was saved, `FALSE` otherwise
}
\description{
Downloads a text file from Google Drive and saves it as a local
  file if the local file does not exist or differs from file in Google Drive.
}
