% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{download_file}
\alias{download_file}
\title{Downloads from Google Docs}
\usage{
download_file(
  file,
  gfile = NULL,
  gpath = "trackdown",
  shared_drive = NULL,
  rm_gcomments = FALSE
)
}
\arguments{
\item{file}{character. The path of a local `.Rmd` or `.Rnw` file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{gpath}{character. (Sub)directory in My Drive or a shared drive
(optional). By default files are uploaded in the folder "trackdown". To
specify another folder the full path is required (e.g., 
"trackdown/my_folder"). Use \code{NULL} to upload directly at the root
level, although it is not recommended.}

\item{shared_drive}{character. The name of a Google Drive shared drive
(optional).}

\item{rm_gcomments}{[experimental] logical value indicating whether or not to
remove Google comments.}
}
\value{
`TRUE` if file from Google Drive was saved, `FALSE` otherwise
}
\description{
Download edited version of a file from Google Drive updating the local
version with the new changes.\cr\cr
\emph{Use with caution as local version of the file will be
overwritten!}\cr\cr 
To know more about \code{trackdown} workflow and features see
\code{\link{trackdown-package}} help page.
}
