% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop quietly}
\usage{
stop_quietly(text = NULL)
}
\arguments{
\item{text}{a string indicating the message to display}
}
\description{
Stop a function without throwing an error.
Function adapted from https://stackoverflow.com/a/42945293/12481476
}
