% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdown_auth.R
\name{trackdown_auth_configure}
\alias{trackdown_auth_configure}
\alias{trackdown_oauth_app}
\title{Edit and view auth configuration}
\usage{
trackdown_auth_configure(app, path)

trackdown_oauth_app()
}
\arguments{
\item{app}{OAuth app, in the sense of [httr::oauth_app()].}

\item{path}{JSON downloaded from Google Cloud Platform Console, containing a
client id (aka key) and secret, in one of the forms supported for the \code{txt}
argument of \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}
}
\value{
* `trackdown_auth_configure()`: An object of R6 class
    [gargle::AuthState], invisibly.
  * `trackdown_oauth_app()`: the current user-configured
    [httr::oauth_app()].
}
\description{
These functions give more control over and visibility into the auth
configuration than [trackdown_auth()] does. `trackdown_auth_configure()`
lets the user specify their own:
  * OAuth app, which is used when obtaining a user token.
See the vignette
[How to get your own API credentials](https://gargle.r-lib.org/articles/get-api-credentials.html)
for more.
If the user does not configure these settings, internal defaults
are used.
`trackdown_oauth_app()` retrieves the currently configured OAuth app.
}
\examples{
# see the current user-configured OAuth app (probaby `NULL`)
trackdown_oauth_app()

if (require(httr)) {

  # store current state, so we can restore
  original_app <- trackdown_oauth_app()

  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  trackdown_auth_configure(app = google_app)

  # confirm current app
  trackdown_oauth_app()
  
  # restore original state
  trackdown_auth_configure(app = original_app)
  trackdown_oauth_app()
}

\dontrun{
# bring your own app via JSON downloaded from Google Developers Console
trackdown_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}

}
\seealso{
Other auth functions: 
\code{\link{trackdown_auth}()},
\code{\link{trackdown_deauth}()}
}
\concept{auth functions}
