% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{update_file}
\alias{update_file}
\title{Updates file in Google Drive}
\usage{
update_file(
  file,
  gfile = NULL,
  path = "rmdrive",
  team_drive = NULL,
  hide_code = FALSE,
  upload_output = FALSE
)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).
By default files are uploaded in the folder "rmdrive". To specify another
folder the full path is required (e.g., "rmdrive/my_folder"). Use
\code{NULL} to upload directly at the root level, although it is not
recommended.}

\item{team_drive}{character. The name of a Google Team Drive (optional).}

\item{hide_code}{logical value indicating whether to remove code from the
text document (chunks and header). Placeholders of  type "[[chunk-<name>]]" are
displayed instead.}

\item{upload_output}{default NULL, specify the path to
the output to upload together with the other file. PDF are directly
uploaded, HTML are first converted into PDF if Chrome is available or as
HTML.}
}
\description{
Replaces the content of an existing file in Google Drive with the contents of
  a local file.
}
\details{
\emph{Use with caution as tracked changes in the Google Drive file will be lost!}
}
