% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{update_file}
\alias{update_file}
\title{Updates file in Google Drive}
\usage{
update_file(
  file,
  gfile = NULL,
  gpath = "trackdown",
  shared_drive = NULL,
  hide_code = FALSE,
  path_output = NULL,
  rich_text = TRUE,
  rich_text_par = NULL,
  force = FALSE,
  open = rlang::is_interactive()
)
}
\arguments{
\item{file}{character. The path of a local \code{.Rmd}, Quarto, or \code{.Rnw} file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{gpath}{character. (Sub)directory in My Drive or a shared drive
(optional). By default files are uploaded in the folder "trackdown". To
specify another folder the full path is required (e.g.,
"trackdown/my_folder"). Use \code{NULL} to upload directly at the root
level, although it is not recommended.}

\item{shared_drive}{character. The name of a Google Drive shared drive
(optional).}

\item{hide_code}{logical value indicating whether to remove code from the
text document (chunks and header). Placeholders of type \code{"[[chunk-<name>]]"}
are displayed instead.}

\item{path_output}{default \code{NULL}, specify the path to the output to
upload together with the other file. PDF are directly uploaded, HTML can be
first converted into PDF if package \code{pagedown} and Chrome are
available.}

\item{rich_text}{(experimental) logical value (default is \code{TRUE})
indicating whether to upload to Google Docs a rich document (i.e.,
important text that should not be changed is highlighted). See “Rich Text”
in details section.}

\item{rich_text_par}{(experimental) argument used to pass a list with custom
settings for rich_text. See “Rich Text” in details section.}

\item{force}{logical value indicating whether to skip confirm check by user
(default is \code{FALSE}).}

\item{open}{logical value indicating whether to open the created document
in a browser (default is \code{TRUE} in interactive sessions).}
}
\value{
a dribble of the uploaded file (and output if specified).
}
\description{
Replaces the content of an existing file in Google Drive with the contents of
a local file. It is also possible to update (or upload if not already
present) the output (pdf or html) of the file specifying the
\code{path_output} argument. In case of html files, if \code{pagedown}
package and Chrome are available, users can decide to upload a pdf version of
the html file.\cr\cr
\emph{Use with caution as tracked changes in the Google Drive file will be
lost!}\cr\cr
To know more about \code{trackdown} workflow and features see
\code{\link[=trackdown-package]{trackdown-package()}} help page.
}
\details{
\strong{Rich Text (experimental)}

The \code{rich_text} option (default is \code{TRUE}) allows to upload a rich
document to Google Docs. Important text that should not be changed is
highlighted. This includes, added instructions at the top of the document,
placeholders hiding the code, header of the document (YAML header or LaTeX
preamble), code chunks, and in-line code.

Default colour is opaque yellow. You can customize the colour specifying the
\code{rgb_color} option in the \code{rich_text_par} argument. The
\code{rgb_color} has to be a list with elements \code{red}, \code{green}, and \code{blue}.
Each element has to be a numeric value between 0 and 1. See example below.
}
\examples{

\dontrun{

# Change default color to opaque light-blue
update_file(file = "path-to/my-file", rich_text = TRUE,
            rich_text_par = list(rgb_color = list(red = 102/255,
                                                  green = 204/255,
                                                  blue = 255/255)))
}

}
