% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{update_rmd}
\alias{update_rmd}
\title{Updates raw rmarkdown for collaborative editing on Google Docs}
\usage{
update_rmd(file, gfile)
}
\arguments{
\item{file}{The name (without file extension) of an .Rmd you want to upload to Google Docs}

\item{gfile}{The name (current or desired) of a Google Docs file (must be unique in your google docs account)}
}
\description{
Updates the content of an existing Google Docs file with the name gfile
}
\examples{
update_rmd(file = "testingfile", gfile = "testingfile-1234")
}
