% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{update_rmd}
\alias{update_rmd}
\title{Updates `.Rmd` file in Google Drive}
\usage{
update_rmd(
  file,
  gfile = basename(file),
  path = "rmdrive",
  team_drive = NULL,
  hide_chunks = FALSE,
  upload_report = FALSE
)
}
\arguments{
\item{file}{character. The path (without file extension) of a local `.Rmd`
file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{path}{character. (Sub)directory in My Drive or a Team Drive (optional).
By default files are uploaded in the folder "rmdrive". To specify another
folder the full path is required (e.g., "rmdrive/my_folder"). Use
\code{NULL} to upload directly at the root level, although it is not
recommended.}

\item{team_drive}{character. The name of a Google Team Drive (optional).}

\item{hide_chunks}{logical value indicating whether to remove code chunks
from the text document. Placeholders of  type "[[chunck_<name>]]" are
displayed instead.}

\item{upload_report}{logical value indicating whether to upload an
additional pdf file with the chunks output (e.g., figures and tables). Note
that this require the time to compile the document.}
}
\description{
Replaces the content of an existing file in Google Drive with the contents of
  a local `.Rmd` file.
}
\details{
\emph{Use with caution as tracked changes in the Google Drive file will be lost!}
}
