% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manager.R
\name{upload_file}
\alias{upload_file}
\title{Upload file to Google Drive for collaborative writing and editing}
\usage{
upload_file(
  file,
  gfile = NULL,
  gpath = "trackdown",
  shared_drive = NULL,
  hide_code = FALSE,
  path_output = NULL,
  rich_text = TRUE,
  rich_text_par = NULL,
  force = FALSE
)
}
\arguments{
\item{file}{character. The path of a local \code{.Rmd}, Quarto, or \code{.Rnw} file.}

\item{gfile}{character. The name of a Google Drive file (defaults to local
file name).}

\item{gpath}{character. (Sub)directory in My Drive or a shared drive
(optional). By default files are uploaded in the folder "trackdown". To
specify another folder the full path is required (e.g.,
"trackdown/my_folder"). Use \code{NULL} to upload directly at the root
level, although it is not recommended.}

\item{shared_drive}{character. The name of a Google Drive shared drive
(optional).}

\item{hide_code}{logical value indicating whether to remove code from the
text document (chunks and header). Placeholders of type \code{"[[chunk-<name>]]"}
are displayed instead.}

\item{path_output}{default \code{NULL}, specify the path to the output to
upload together with the other file. PDF are directly uploaded, HTML can be
first converted into PDF if package \code{pagedown} and Chrome are
available.}

\item{rich_text}{\link{experimental} logical value (default is \code{TRUE})
indicating whether to upload to Google Docs a rich document (i.e.,
important text that should not be changed is highlighted). See “Rich Text”
in details section.}

\item{rich_text_par}{\link{experimental} argument used to pass a list with custom
settings for rich_text. See “Rich Text” in details section.}

\item{force}{logical value indicating whether to skip confirm check by user
(default is \code{FALSE}).}
}
\value{
a dribble of the uploaded file (and output if specified).
}
\description{
Uploads a local file to Google Drive as a plain text document. Will only
upload the file if it doesn't already exist in the chosen location. By
default files are uploaded in the folder "trackdown", if is not available on
Google Drive, permission to create it is required to the user. To update an
already existing file see \code{\link[=update_file]{update_file()}}. It is also possible to
upload the output (pdf or html) of the file specifying the \code{path_output}
argument. In case of html files, if \code{pagedown} package and Chrome are
available, users can decide to upload a pdf version of the html file.\cr\cr
To know more about \code{trackdown} workflow and features see
\code{\link[=trackdown-package]{trackdown-package()}} help page.
}
\details{
\strong{Rich Text \link{experimental}}

The \code{rich_text} option (default is \code{TRUE}) allows to upload a rich
document to Google Docs. Important text that should not be changed is
highlighted. This includes, added instructions at the top of the document,
placeholders hiding the code, header of the document (YAML header or LaTeX
preamble), code chunks, and in-line code.

Default colour is opaque yellow. You can customize the colour specifying the
\code{rgb_color} option in the \code{rich_text_par} argument. The
\code{rgb_color} has to be a list with elements \code{red}, \code{green}, and \code{blue}.
Each element has to be a numeric value between 0 and 1. See example below.
}
\examples{

\dontrun{

# Change default color to opaque light-blue
upload_file(file = "path-to/my-file", rich_text = TRUE,
            rich_text_par = list(rgb_color = list(red = 102/255,
                                                  green = 204/255,
                                                  blue = 255/255)))
}

}
