% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd.R
\name{upload_rmd}
\alias{upload_rmd}
\title{Uploads raw rmarkdown for collaborative editing on Google Docs}
\usage{
upload_rmd(file, gfile)
}
\arguments{
\item{file}{The name (without file extension) of an .Rmd you want to upload to Google Docs}

\item{gfile}{The name (current or desired) of a Google Docs file (must be unique in your google docs account)}
}
\description{
Allows upload of an .Rmd file as to Google Docs as a plain text document
}
\examples{
upload_rmd(file = "testingfile", gfile = "testingfile-1234")
}
